+++ packages/libraries/client/tests/integration.spec.ts
@@ -1,12 +1,11 @@
 import { createServer } from 'node:http';
 import { AddressInfo } from 'node:net';
-
-/* eslint-disable-next-line import/no-extraneous-dependencies */
-import { ApolloServerBase } from 'apollo-server-core';
 import axios from 'axios';
 
 /* eslint-disable-next-line import/no-extraneous-dependencies */
 import { createSchema, createYoga } from 'graphql-yoga';
+// eslint-disable-next-line import/no-extraneous-dependencies
+import { ApolloServer } from '@apollo/server';
 import { createHive, hiveApollo, useHive } from '../src';
 import { waitFor } from './test-utils';
 
@@ -40,8 +39,8 @@ function handleProcess() {
 
 test('GraphQL Yoga - should not interrupt the process', async () => {
   const logger = {
-    error: jest.fn(),
-    info: jest.fn(),
+    error: vi.fn(),
+    info: vi.fn(),
   };
   const clean = handleProcess();
   const hive = createHive({
@@ -106,11 +105,11 @@ test('GraphQL Yoga - should not interrupt the process', async () => {
 
 test('Apollo Server - should not interrupt the process', async () => {
   const logger = {
-    error: jest.fn(),
-    info: jest.fn(),
+    error: vi.fn(),
+    info: vi.fn(),
   };
   const clean = handleProcess();
-  const apollo = new ApolloServerBase({
+  const apollo = new ApolloServer({
     typeDefs,
     resolvers,
     plugins: [
@@ -136,7 +135,6 @@ test('Apollo Server - should not interrupt the process', async () => {
     ],
   });
 
-  await apollo.start();
   await apollo.executeOperation({
     query: /* GraphQL */ `
       {