+++ packages/services/usage/src/index.ts
@@ -185,7 +185,7 @@ async function main() {
     if (env.prometheus) {
       await startMetrics(env.prometheus.labels.instance ?? undefined);
     }
-    await server.listen(env.http.port, '0.0.0.0');
+    await server.listen(env.http.port, '::');
     await start();
   } catch (error) {
     server.log.fatal(error);