+++ packages/web/app/src/components/organization/Usage.tsx
@@ -1,5 +1,4 @@
 import React from 'react';
-import 'twin.macro';
 import { useQuery } from 'urql';
 import {
   OrganizationFieldsFragment,
@@ -30,7 +29,7 @@ export const OrganizationUsageEstimationView: React.FC<{
 
   return (
     <>
-      <div tw="top-7 right-4">
+      <div className="top-7 right-4">
         <DataWrapper query={query}>
           {result => (
             <TableContainer>
@@ -55,7 +54,7 @@ export const OrganizationUsageEstimationView: React.FC<{
                         value={result.data.usageEstimation.org.operations}
                         size={10}
                         max={organization.rateLimit.operations}
-                        tw="justify-end"
+                        className="justify-end"
                       />
                     </Td>
                   </Tr>