+++ packages/web/app/src/components/target/Switcher.tsx
@@ -1,6 +1,5 @@
 import React from 'react';
 import { VscChevronDown } from 'react-icons/vsc';
-import 'twin.macro';
 import { useQuery } from 'urql';
 import { TargetsDocument } from '@/graphql';
 import { useRouteSelector } from '@/lib/hooks';
@@ -40,7 +39,7 @@ export const TargetSwitcher: React.FC<{
         as={Button}
         rightIcon={<VscChevronDown />}
         variant="ghost"
-        tw="font-normal"
+        className="font-normal"
       >
         {currentTarget.name}
       </MenuButton>