+++ packages/web/app/src/components/target/operations/Filters.tsx
@@ -2,7 +2,6 @@ import React, { ComponentType } from 'react';
 import { VscChevronDown, VscChromeClose } from 'react-icons/vsc';
 import AutoSizer from 'react-virtualized-auto-sizer';
 import { FixedSizeList, ListChildComponentProps } from 'react-window';
-import 'twin.macro';
 import { useQuery } from 'urql';
 import { useDebouncedCallback } from 'use-debounce';
 import { Spinner } from '@/components/common/Spinner';
@@ -111,7 +110,7 @@ const OperationsFilter: React.FC<{
         <DrawerHeader bgColor="gray.900">Filter by operation</DrawerHeader>
         <DrawerCloseButton />
         <DrawerBody bgColor="gray.900">
-          <div tw="flex flex-col h-full space-y-3">
+          <div className="flex flex-col h-full space-y-3">
             <InputGroup>
               <Input
                 pr="3rem"
@@ -133,7 +132,7 @@ const OperationsFilter: React.FC<{
                 />
               </InputRightElement>
             </InputGroup>
-            <div tw="flex flex-row justify-between items-center">
+            <div className="flex flex-row justify-between items-center">
               <div>
                 <Button variant="link" size="xs" onClick={selectAll}>
                   All
@@ -142,7 +141,7 @@ const OperationsFilter: React.FC<{
                   None
                 </Button>
               </div>
-              <div tw="flex flex-row">
+              <div className="flex flex-row">
                 <Button variant="ghost" size="sm" onClick={selectAll}>
                   Reset
                 </Button>
@@ -159,7 +158,7 @@ const OperationsFilter: React.FC<{
                 </Button>
               </div>
             </div>
-            <div tw="pl-1 flex-grow">
+            <div className="pl-1 grow">
               <AutoSizer>
                 {({ height, width }) => (
                   <FixedSizeList
@@ -241,13 +240,13 @@ const OperationRow: React.FC<{
   }, [onSelect, operation.operationHash, selected]);
 
   return (
-    <div style={style} tw="flex flex-row space-x-4 items-center">
+    <div style={style} className="flex flex-row space-x-4 items-center">
       <Checkbox colorScheme="primary" isChecked={selected} onChange={change} />
-      <div tw="flex flex-grow flex-row items-center cursor-pointer">
-        <button tw="flex-grow overflow-ellipsis overflow-hidden whitespace-nowrap" onClick={change}>
+      <div className="flex grow flex-row items-center cursor-pointer">
+        <button className="grow text-ellipsis overflow-hidden whitespace-nowrap" onClick={change}>
           {operation.name}
         </button>
-        <button tw="width[75px] flex-shrink-0 text-right text-gray-500" onClick={change}>
+        <button className="w-[75px] shrink-0 text-right text-gray-500" onClick={change}>
           {requests}
         </button>
       </div>
@@ -274,7 +273,9 @@ export const OperationsFilterTrigger: React.FC<{
         bgColor="whiteAlpha.50"
         _hover={{ bgColor: 'whiteAlpha.100' }}
       >
-        <span tw="font-normal">Operations ({selected?.length ? selected.length : 'all'})</span>
+        <span className="font-normal">
+          Operations ({selected?.length ? selected.length : 'all'})
+        </span>
       </Button>
       <OperationsFilterContainer
         isOpen={isOpen}