+++ scripts/seed-local-env.js
@@ -13,7 +13,7 @@ async function main() {
         ? 'https://app.staging.graphql-hive.com/registry'
         : process.env.DEV
         ? 'https://app.dev.graphql-hive.com/registry'
-        : 'http://localhost:4000/graphql',
+        : 'http://localhost:3001/graphql',
       author: 'Hive Seed Script',
       commit: '1',
     },
@@ -67,7 +67,7 @@ async function main() {
     () => {
       for (let i = 0; i < operationsPerBatch; i++) {
         const randNumber = Math.random() * 100;
-        console.log(`Reporting usage query...`);
+        console.log('Reporting usage query...');
 
         const done = hiveInstance.collectUsage({
           document: randNumber > 50 ? query1 : query2,